+++ R/swapInfo.R
@@ -1,8 +1,9 @@
 #' Commodity Swap details to learn their pricing
 #' @description Returns dataframe required to price a WTI averaging instrument based on first line settlements.
 #' @param date Character date as of which you want to extract daily settlement and forward values. `character`
-#' @param feeds Feeds for Morningstar getCurve() and getPrice(). `character`
+#' @param feed Feeds for Morningstar getCurve() and getPrice(). `character`
 #' @param exchange Exchange code in data(holidaysOil). Defaults to "nymex". `character`
+#' @param ticker  Nymex contract code. `character`
 #' @param contract Contract code in data(expiry_table). sort(unique(expiry_table$cmdty)) for options. `character`
 #' @param iuser Morningstar user name as character - sourced locally in examples. `character`
 #' @param ipassword Morningstar user password as character - sourced locally in examples. `character`
@@ -12,31 +13,24 @@
 #' @author Philippe Cote
 #' @examples
 #' \dontrun{
-#' feeds <- dplyr::tibble(
-#'   feed = c(
-#'     "Crb_Futures_Price_Volume_And_Open_Interest",
-#'     "CME_NymexFutures_EOD_continuous"
-#'   ),
-#'   ticker = c("CL", "CL_001_Month")
-#' )
 #' swapInfo(
-#'   date = "2020-05-06", feeds = feeds, contract = "cmewti", exchange = "nymex",
+#'   date = "2020-05-06", feed = "CME_NymexFutures_EOD_continuous",
+#'   ticker = "CL",
+#'   contract = "cmewti", exchange = "nymex",
 #'   iuser = "x@xyz.com", ipassword = "pass", output = "all"
 #' )
 #' }
 #'
-swapInfo <- function(date = "2020-05-06",
-                     feeds = dplyr::tibble(
-                       feed = c("Crb_Futures_Price_Volume_And_Open_Interest", "CME_NymexFutures_EOD_continuous"),
-                       ticker = c("CL", "CL_001_Month")
-                     ),
+swapInfo <- function(date = "2023-08-24",
+                     feed = "CME_NymexFutures_EOD_continuous",
+                     ticker = "CL",
                      contract = "cmewti", exchange = "nymex",
                      iuser = "x@xyz.com", ipassword = "pass", output = "all") {
   fig.title <- paste("Swap Pricing:", contract, "Futures Curve as of", date)
 
   wti <- getCurve(
-    feed = feeds$feed[1], contract = feeds$ticker[1],
-    date = date, fields = c("Close"), iuser = iuser, ipassword = ipassword
+    feed = feed, contract = ticker,
+    date = date, fields = c("settlement_price"), iuser = iuser, ipassword = ipassword
   )
 
   to <- as.Date(date) + months(4)
@@ -46,11 +40,12 @@ swapInfo <- function(date = "2020-05-06",
   bizDays <- bizDays[!(weekdays(bizDays) %in% c("Saturday", "Sunday"))]
 
   curve <- dplyr::tibble(bizDays = bizDays) %>%
-    dplyr::left_join(wti %>% dplyr::select(bizDays = expirationDate, fut.contract = code, price = Close), by = c("bizDays")) %>%
-    tidyr::fill(price, fut.contract, .direction = c("up"))
+    dplyr::left_join(wti %>% dplyr::select(bizDays = expirationDate, fut.contract = code, price = Settlement), by = c("bizDays")) %>%
+    tidyr::fill(price, fut.contract, .direction = c("up")) %>%
+    tidyr::drop_na()
 
   hist <- RTL::getPrice(
-    feed = feeds$feed[2], contract = feeds$ticker[2],
+    feed = feed, contract = paste0(ticker,"_001_MONTH"),
     from = as.character(lubridate::rollback(as.Date(date), roll_to_first = T)),
     iuser = iuser, ipassword = ipassword
   ) %>%